Fixed bug's
