+++ js/length-checker.js
@@ -1,0 +1,6 @@
+const getCorrectLength = (actualLength, maxLength) => {
+  const result = maxLength >= actualLength;
+  return result;
+};
+
+export {getCorrectLength};
