+++ js/main.js
@@ -1,9 +1,10 @@
-import './template-photos.js';
+import {createPhotoObjects} from './blank-photos.js';
+import {getCorrectLength} from './length-checker.js';
+import {renderPhotos} from './template-photos.js';
 
-const getCorrectLength = (actualLength, maxLength) => {
-  const result = maxLength >= actualLength;
-  return result;
-};
+const MAX_PHOTOS = 25;
+const photos = createPhotoObjects(MAX_PHOTOS);
 
 
 getCorrectLength(150, 140);
+renderPhotos(photos);
